# Refactor LTable to improve performance