Improve conversion speed between lua value and go value by reduce memory allocation。

### Changes proposed in this pull request:

append lTableDictValue{value LValue, keyIndex int}.
change LTable would contains unique dict map[LValue]lTableDictValue.
change LTable#keys as []*LValue to reduce memory allocation on RawSetString.
append LTable#DictionaryLen() (use to initial go map cap).
append some convert functions between LValue and go value.

### Benchmark testing:
- Load the [28kb json](https://raw.githubusercontent.com/valyala/fastjson/master/testdata/large.json) decode as map[string]interface{}, convert to LTable and then convert back.
- original
BenchmarkToLValue-4   	    2157	    559233 ns/op	  272908 B/op	    3799 allocs/op
BenchmarkToGValue-4   	    3924	    270173 ns/op	  126799 B/op	    2025 allocs/op
- changed
BenchmarkToLValue-4   	    2704	    377055 ns/op	  187883 B/op	    3459 allocs/op
BenchmarkToGValue-4   	    6284	    169170 ns/op	   82971 B/op	     981 allocs/op

I would charge test code, if the changing would be accepted.