@tul 2020-12-10
I think that maybe this change:

" change LTable#keys as []*LValue to reduce memory allocation on RawSetString."

May not be doing what's wanted.

`LValue` is an interface type, and as such is modelled behind the scenes as:

```
type iface struct {
        tab  *itab
        data unsafe.Pointer
}
```

So a slice of LValues `[]LValue` is modelled in memory as:

```
*itab
unsafe.Pointer
*itab
unsafe.Pointer
*itab
*unsafe.Pointer
....
```

I think changing the type to be `[]*LValue` will actually change the slice to be an array of pointers to new allocated `iface` structures on the heap, which will then point to the actual value. i.e.

```
*iface
*iface
*iface
....

These `iface` elements will be heap allocated. In other words, it will introduce an additional redirection and also potentially spread the values around the heap so that linear iteration of the slice may potentially incur additional cache misses.

I did an investigation of slices of `interface{}` a couple of years back and found it to be like this, I did a write up here: https://tul.github.io/2018/07/23/go-interfaces-deep-dive.html

So I actually think this particular aspect of the change may be detrimental to performance.